Added win32 implementation of native_midi_setvolume(), compliments of  "Java the Hut" on the 3DRealms forums:
   http://forums.3drealms.com/ubb/ultimatebb.php?ubb=get_topic;f=24;t=000014

Java the Hut
  posted 01-01-2003 10:44 PM

I found a missing feature in the SDL_mixer library. The volume control does
 not work for the native Win32 midi feature.

The void native_midi_setvolume(int volume) function in the native_midi_win32.c
 file has no code, so changing the music volume does nothing.

This code adds volume control for the midi music in the Win32 ROTT. You will
 have to download the SDL_mixer code and compile it yourself.

  [...code snipped...]

It is kind of cool to have a working Win32 ROTT with sound and music.


git-svn-id: http://svn.libsdl.org/trunk@884 c70aab31-4412-0410-b14c-859654838e24