ate: Mon, 4 Aug 2003 19:10:14 +0300 From: "Mike Gorchak" Subject: SDL/QNX forgotten fixes
Just found, that one chunk for configure.in has not been applied and one for\
gotten by me :) Please add it to CVS.

P.S. Now QNX6.2.1 have the dlopen in libc and in the libltdl. So libc's vari\
ant more prefered, and ltdl rest for compatibility. In this small patch I wa\
s added proper checks.


git-svn-id: http://svn.libsdl.org/trunk@1039 c70aab31-4412-0410-b14c-859654838e24