Date: Sat, 16 Aug 2003 16:22:56 +0300 From: "Mike Gorchak" Subject: Package building for QNX6
I'm just completed the package description file for QNX6 - qpg, it is like a .spec files for Linux. Please place SDL.qpg.in file in the root of the project, where .spec file is placed. And sdl12qpg.diff - just adding the SDL.qpg.in. The same for the SDL_image. I'm planning to add .qpg files creation for all SDL* projects.


git-svn-id: http://svn.libsdl.org/trunk@1086 c70aab31-4412-0410-b14c-859654838e24