Date: Sat, 24 Jan 2004 14:49:58 +0100 From: Stephane Marchesin Subject: Re: [SDL] 3DNow! detection on a P4 system
Well... I found another bug : 3dnow was detected on P2/P3 cpus. So I
took one more look at the code and found a mistake in
CPU_getCPUIDFeaturesExt :
The condition for having extended cpuinfo is that when we query extended
cpuinfo we get a result >=  0x80000001. So we must exit if eax < 0x80000001.
The attached patch does that.


git-svn-id: http://svn.libsdl.org/trunk@1210 c70aab31-4412-0410-b14c-859654838e24