Date: Sun, 15 Feb 2004 23:51:06 +0100 From: Max Horn Subject: Some more things re: SDL website
Some minor nitpicks:

* it's a good idea to always specify width/height for <img> -> that
speeds up rendering a lot, because the browser can now do the layout
before the pictures are loaded. The attached files fix this, at least
for the pics in header.inc

* absolute paths vs. relative: I changed image refs from "/images" to
"images", which makes it easier to test the site on other machines.
However, I dunno if this actually works on the SDL web server; but
undoing this change in header.inc should be trivial

I also corrected various validation errors, and added a CSS stylesheet
file. You'll find all of this in the attached file -- maybe it's of
some use for you. Also, I recommend
http://www.alistapart.com/articles/taminglists/ for a nice introduction
to using CSS for styling lists (even using nice hover effects).


git-svn-id: http://svn.libsdl.org/trunk@1265 c70aab31-4412-0410-b14c-859654838e24