Date: Mon, 16 Feb 2004 19:21:51 +0100 From: Max Horn Subject: Re: Fix for OSX bug
The problem is that our actual client area is realized as a
subview of the "content view" of the window. That's normal so far. The
odd part is that, while our subview starts out filling it parent 100%,
upon resizes it gets changed to something smaller....

Anyway, here is a revised patch which simply enforces the size of
window_view by calling setFrameSize.


git-svn-id: http://svn.libsdl.org/trunk@1277 c70aab31-4412-0410-b14c-859654838e24