Date: Tue, 02 Mar 2004 00:10:30 +0100 From: St���han Koch Subject: Tiny annoyance with the new site design
I'm not sure where to comment about the site, so the most straight-
forward thing for me to do was to bonk the feedback link. :)
The new site design is lovely, I must say. There's one nasty bug however
that affects high resolutions: the site doesn't extend all the way to
the bottom.
My suggestion is, get rid of the table's bgcolor and table cell's
background properties here:

> <!-- Main -->
> <table cellpadding="0" cellspacing="0" border="0" width="100%" bgcolor="#B\
0D7FA">
> <tr>
>
>       <!-- Side Bar-->
>       <td background="/images/barbg.png" valign="top">

And make the body element in the stylesheet file read:
> body
> {
>       margin: 0px;
>       background: #B0D7FA url('/images/barbg.png') top left repeat-y;
> /*    font-family: Verdana, Tahoma, Arial, Helvetica, sans-serif; */
> }

That should fix it, I think.
Once again, great work on the site design to whoever did it. :)


git-svn-id: http://svn.libsdl.org/trunk@1315 c70aab31-4412-0410-b14c-859654838e24