[SDL] [PATCH] fix FB_VideoQuit for ia64 Jesse Barnes Fri Jan 14 11:13:54 PST 2005
I noticed that on my ia64 machine when SDL_Quit was called, the machine would
hang in weird ways.  It turned out to be caused by a machine check in the
memset() call near the top of FB_VideoQuit.  Generally memset shouldn't be
used on I/O regions like the framebuffer or other I/O device memory (the
Linux kernel has special routines for dealing with I/O memory), so I changed
the #ifdef __powerpc__ to #if defined(__powerpc__) || defined(__ia64__) and
SDL_Quit now works properly.  The change should probably be made
unconditional though since doing one byte at a time is the only way to be
portable, afaik.


git-svn-id: http://svn.libsdl.org/trunk@1529 c70aab31-4412-0410-b14c-859654838e24