Date: Fri, 22 Apr 2005 21:07:49 -0600 From: Tyler Montbriand Subject: [SDL] SDL_mixer mikmod patch
When SDL_mixer calls MikMod_RegisterAllDrivers using it's internally
compiled-in audio drivers, the only driver it registers is the nosound audio
driver.  When it's using an external mikmod library, it initializes any and
all drivers it has available, such as OSS, ESD, and so forth.  Because of the
way SDL_mixer uses mikmod directly, this will often work, but will do silly
things like open the audio device twice -- once by SDL, once by MikMod.
Worst case it will hang hard as MikMod attempts to open a busy audio device.

All it needs is the nosound driver for it to work, not ALL mikmod drivers.
The attached patch causes it to only initialize that one when using an
external mikmod library.


git-svn-id: http://svn.libsdl.org/trunk@1555 c70aab31-4412-0410-b14c-859654838e24