(Compliments of Eric Wing...)
Hi Ryan,
     It was really nice meeting you at WWDC.

     Attached is the file containing the fix for the SDL_mixer issue
we stumbled upon. I realized that I took this file out of stable
instead of CVS so I wasn't sure if a diff would be very useful.

     I tried to document the problem. Just do a search for "Apple" in
the file. Basically, the Quicktime people screwed up and they use two
entirely different C types depending on your platform's endian-ness.
So you have to case the silly thing. (So technically, I think this is
actually a porting issue and not a gcc 4.0 issue, but it's not our
fault.)

Thanks,
Eric


git-svn-id: http://svn.libsdl.org/trunk@1577 c70aab31-4412-0410-b14c-859654838e24