te: Thu, 17 Apr 2003 11:25:26 -0700 From: "Jim" Subject: [SDL] Frame buffer patches...
Would seem that GCC 3.2.2 promotes all signed values to unsigned if any single vlaue is unsigned in an expression...

for instance when calculating an elo touch position....evertyhign is an (int) except for cach_vinfo.xres which is unsigned.  THerefore if y
ou reverse the SDL_ELO_MIN_X and SDL_ELO_MAX_X values the resulging negative value pegs the expression to it's max.  Attached it the accumu
lated patch to typecast the unsigned values to (int)

- *dx = (cache_vinfo.xres - (cache_vinfo.xres * (input_x - ELO_MIN_X)) / width);

+ *dx = ((int)cache_vinfo.xres - ((int)cache_vinfo.xres * (input_x - ELO_MIN_X)) / width);

and also to provide quite failure to operations which have the screen currently locked...

The touch screen I'm using the original values were exactly opposite of any position I touched on the screen - evaluating the math the expr
ession SHOULD have handled reversing the min and max values - and after casting the unsigned .xres and .yres to a signed int - worked well.
..

Jim


git-svn-id: http://svn.libsdl.org/trunk@1792 c70aab31-4412-0410-b14c-859654838e24