Date: Fri, 24 Dec 2004 16:11:30 +0100 From: Gaetan de Menten Subject: Re: [SDL] SDL_mixer and implicit looping of module music files
Maybe it will be clearer with this information copy-pasted from
Mikmod's documentation:

BOOL wrap
    If nonzero, module wraps to its restart position when it is
finished, to play continuously. Default value is zero (play only
once).

BOOL loop
    If nonzero, all in-module loops are processed; otherwise, backward
loops which decrease the current position are not processed (i.e. only
forward loops, and backward loops in the same pattern, are processed).
This ensures that the module never loops endlessly. The default value
is 1 (all loops are processed).

and this is from SDL_mixer's code:

music->data.module->wrap    = 0;
music->data.module->loop    = 0;

What I don't like is the loop = 0 (note it is not the default value in
mikmod) and that there is no way to change it, the wrap = 0 is fine
with me.

-Gaetan.


git-svn-id: http://svn.libsdl.org/trunk@1818 c70aab31-4412-0410-b14c-859654838e24