Date: Mon, 13 Mar 2006 14:38:36 +1100 From: Alison Winters Subject: SDL RTF, charsets
What i'd like to do is store the "charset" value for the fonts used in the RTF
document.  This will allow client apps (SDL as well as our own app) to parse
the charset and pass characters through a translation table so that they will
display appropriately in the actual font being used.  This is important
because Microsoft Word makes RTF files that use (for example) the Greek
charset to display Alpha, Beta, Gamma etc. instead of using the values in the
ANSI charset.

Here are the diffs for the four public domain RTF files.  Let me know if you
feel it makes sense for the way SDL does things.  There is one more change
required, however - a change to the function RTF_CreateFont must be made in
the SDL_rtfreadr.c module.  There is now an extra parameter called charset
being passed in.  Initially you could just ignore the extra parameter and
everything will work as it used to.

There is a FIXME listed in SDL_rtf.h that covers the CreateFont function of
the RTF_FontEngine object, however.  This could be altered to add the
charset.  After that it starts getting a bit SDL-specific, however - the
programmer will have to do the rendering work himself.  I'm not sure if you
can specify a Windows charset in a "TTF_Font" object; if not you'll have to do
some clever stuff using conversion tables from Windows character sets/code
pages to Unicode.

For our product we just implemented Greek, Turkish and Baltic, but in a large
project like SDL you're probably using libintl or something anyway, so there
may be auto-conversion functions you can call.  If it is going to have to be
self-implemented, here is where i got the conversion tables from:
http://dev.icu-project.org/cgi-bin/viewcvs.cgi/charset/data/  It's part of
some IBM project or other and lists pretty much everything you need in XML - i
just did some playing around with sed and sort to get it in a format i could
paste direct into the C code.  Let me know if you need any more pointers with
that, though it should be fairly straightforward.

I'd do a bit more for you on the SDL front, but i'm not really 100% familiar
with how you implement internationalization and all that stuff in SDL so i'd
rather it was put on a TODO list for someone who knew what they were doing ;-)

Thanks,
Alison


git-svn-id: http://svn.libsdl.org/trunk@2083 c70aab31-4412-0410-b14c-859654838e24