Fixed bug #191 [I opted to go for a warning, since I often tweak configure.in, but don't want to rebuild the entire project]
One thing that was lost in the switch from automake to the new build system is
that there is now no rule to build configure from configure.in.

IMHO, if configure.in gets changed, then at the very least, the build system
should print out a warning (better, again IMHO, an error) about this fact.
Else, you easily forget about this when modifying configure.in.


git-svn-id: http://svn.libsdl.org/trunk@2180 c70aab31-4412-0410-b14c-859654838e24