From: Torsten Giebl Subject: ALLOCA Patch for SDL-1.2 Newsgroups: gmane.comp.lib.sdl Date: 2006-10-24 00:31:16 GMT
Hello !

As alloca.h is not available on CYGWIN and MinGW32
it should not be checked there or it should be checked
with -mno-cygwin using.

I am a total configure.in newbie, but i found the way
that things are handled in configure.in pretty bad for the
case CYGWIN with MinGW Mode or not.

Maybe this is not possible but i would like to
have a way to detect at the start for example CYGWIN
and then add. ask if configure was called with --enable-cygwin
for example. --enable-cygwin should be optional and disabled
by default as we want to have MinGW Mode by default.

Only if the user wants it he should be able to use it.

Then the whole configure.in would get add. questions
okay the system is CYGWIN with or without MinGW Mode.

The alloca.h thing for example is only
available under CYGWIN ( without MinGW Mode ).

CU



git-svn-id: http://svn.libsdl.org/trunk@2922 c70aab31-4412-0410-b14c-859654838e24