Date: Wed, 23 May 2007 00:39:47 +0300 From: "Eyal Lotem" To: "SDL Mailing List" <sdl@lists.libsdl.org> Subject: [SDL] Bug in GL_SWAP_CONTROL extension name.
The checked extension name is incorrect.
The attached patch fixes it to the correct name.

I don't think that under any setup, this extension name was ever
correct, which is why its not OR'd against the old check.

This fixes the "tear effect" (vsync problems) I had here with nVidia
GLX drivers.

Eyal



git-svn-id: http://svn.libsdl.org/trunk@3021 c70aab31-4412-0410-b14c-859654838e24