Austen Dicken - Tue Feb 26 23:28:27 PST 2008
Ok, here is the patch I made for FLAC support.

I have tested it relatively thoroughly and currently the patch allows:
 1. Pre-loading FLAC files and playing them via LoadWAV
 2. The patch allows for FLAC support in the LoadMUS setting as well as:
     *  Pause / Resume
     *  Volume control
     *  Seeking

I also did a little benchmarking by comparing memory/cpu usage of playmus to
that of mplayer, and the results were very good.  playmus typically took about
half the RAM as mplayer, though that may be attributed to mplayer being a more
"bulky" program.  As such I would say that the two are probably about equal in
efficiency.

Also, it is important to note that, similar to the OGG support currently
built-in, my FLAC patch only supports 16 bit stereo-encoded sound.  Also, it
is only for Native FLAC (standard) and not the derivative, Ogg-FLAC.

I have tried to find a simple way to detect Ogg-FLAC files, as the only
difference between Ogg-FLAC and Native FLAC support is changing the init_
function call, but after digging a little deeper it seems that Ogg-FLAC is
basically FLAC wrapped in an Ogg transport layer, so it would be better to have
a way to read the Ogg transport layer which then reads the inner audio files
according to the proper codec.

But anyway, that's another job for another day!  For now this should provide
Native FLAC support!



git-svn-id: http://svn.libsdl.org/trunk@3557 c70aab31-4412-0410-b14c-859654838e24