My fix for this reported issue...  --ryan.
Subject: 	Bug in SDL_mixer
Date: 	Thu, 6 Mar 2008 16:46:45 +1300
From: 	Karl Hendrikse
To: 	jcatki@jcatki.no-ip.org

I really don't know if I'm sending this to anywhere near the right
person, but I believe I have found a bug in SDL_mixer.

In mixer.c, in the mix_channels function, line 190 should be moved
outside the if statement. This will reset the volume of a channel at the
end of any fade, not just after a fade out.

The way the code works now assumes the volume will always be correct at
the end of a fade in, but in reality it will usually never get there
because the last time the volume gets updated, the fade is not quite
complete (ie. the volume is always up to one frame behind where it would
ideally be). This issue gets worse as the fadein time decreases (since
the jump in volume each iteration is greater).



git-svn-id: http://svn.libsdl.org/trunk@3564 c70aab31-4412-0410-b14c-859654838e24